Fix bug when player moves on tile selection
This fixes a bug where the player block moved to a position based on its
name. This is not correct because we updated a tile block's name to
include the y-sorting information.

Prior to this fix, it was performing a split over the `_`
underscore character and assuming the first two values would be the `x`
and `z` values. However, this is no longer the case as the first value
is now the `y`-value.