Add movement based on camera
Previously, the movement was based on the Input axis being mapped to
player movement based on a fixed coordinate system of the world. This
meant that regardless of the camera rotation, pressing right would make
the player move in the same direction with respect to the map. This
results in an unintuitive control scheme because players may choose to
rotate the camera in any way that they please.

This change thus bases the player's movement on the direction that the
camera is facing. It uses the Camera.main.transform's facing directional
vector and uses that to figure out the movement vector of the player.
For pressing left and right keys (i.e., Input.Axis("Horizontal"), it
rotates the movement vector by 90 degrees abouve the y-axis, and
multiplies it by the respective coefficient. This is also the first time
that I've used Quarternions for such a purpose, so Thank God that it
worked the first time around!